refactor(launch/widgets): split file_browser into responsibility submodules
`file_browser.rs` grew to ~1800 lines and combined filesystem
scanning, sandbox policy, git-repo detection, git URL resolution,
keyboard navigation, a git-prompt state machine, mouse URL hit-test,
rendering, and a large test suite. Split into `file_browser/` as a
directory module:

- `file_browser/mod.rs` — module doc, shared palette, module
  declarations, cross-module re-exports.
- `file_browser/state.rs` — `FileBrowserState`, `FolderEntry`,
  sandbox helpers (`is_within_root`, `canonicalize_or_self`),
  listing helpers (`load_entries`, `has_git_dir`, `is_excluded`),
  navigation (`set_cwd`, `reload`, `navigate_up`).
- `file_browser/input.rs` — keyboard/mouse event handling
  (`handle_key`, `handle_enter`, `commit_or_reject`,
  `maybe_open_url_on_click`).
- `file_browser/render.rs` — `render`, row styling.
- `file_browser/git_prompt.rs` — `GitPromptFocus`,
  `handle_git_prompt_key`, `render_git_prompt`, `resolve_git_url`.

Tests travel with the behavior they exercise. No behavior change —
keyboard behavior, sandbox rules, mouse URL opening, and rendering
are all byte-identical.

Addresses finding #6 of the PR #166 current-branch review.
