--- src/launch/widgets/file_browser.rs
@@ -1,1809 +1,0 @@
-//! Host folder picker — custom directory browser scoped to $HOME.
-//!
-//! Behavior:
-//! - Starts at $HOME.
-//! - Refuses navigation above $HOME (clamps cwd back to root).
-//! - Excludes noisy top-level directories at the $HOME level.
-//! - Rejects $HOME itself and ~/.jackin/* as workspace sources.
-//! - Tags git-repo rows with a trailing ` (git)` suffix in a distinct
-//!   colour so the operator can scan for repos at a glance. Enter on a
-//!   repo row opens a prompt (mount / pick-subdir / cancel) before
-//!   committing or navigating in.
-//!
-//! The browser was originally built on `ratatui-explorer`, but that
-//! crate's `Theme` exposes a single `dir_style` shared by every row —
-//! meaning "colour git repos differently" is impossible. Rewriting in-
-//! house costs ~400 lines and unlocks per-entry styling plus a simpler
-//! keymap (`h/l` / arrows / `s` / `Esc` handled directly instead of
-//! round-tripping through the explorer's event handler).
-
-use std::path::{Path, PathBuf};
-
-use crossterm::event::{KeyCode, KeyEvent};
-use directories::BaseDirs;
-use ratatui::{
-    Frame,
-    layout::Rect,
-    style::{Color, Modifier, Style},
-    text::{Line, Span},
-    widgets::{Block, Borders, Paragraph},
-};
-use tui_widget_list::ListState;
-
-/// Phosphor green — matches jackin's primary colour.
-const PHOSPHOR_GREEN: Color = Color::Rgb(0, 255, 65);
-/// Dimmed phosphor — used for the ` (git)` suffix and italic metadata.
-const PHOSPHOR_DIM: Color = Color::Rgb(0, 140, 30);
-/// Bright white — used for cwd titles + focus highlights.
-const WHITE: Color = Color::Rgb(255, 255, 255);
-/// Sandbox-rejection / error red.
-const DANGER_RED: Color = Color::Rgb(255, 94, 122);
-/// Dark phosphor — block borders, separator glyphs.
-const PHOSPHOR_DARK: Color = Color::Rgb(0, 80, 18);
-
-use super::ModalOutcome;
-
-/// Directories excluded from the listing when browsing $HOME.
-const EXCLUDED: &[&str] = &[
-    "Library",
-    "Applications",
-    "Movies",
-    "Music",
-    "OrbStack",
-    "Pictures",
-];
-
-/// Does `path` contain a `.git` child? Dir (regular clone) OR file
-/// (submodule worktree, `.git` is a file pointing at the real gitdir).
-/// Single `metadata` call per directory entry — no filesystem walk.
-fn has_git_dir(path: &Path) -> bool {
-    let dotgit = path.join(".git");
-    dotgit.is_dir() || dotgit.is_file()
-}
-
-/// Focus target for the in-browser "git-repo row, what now?" prompt.
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub enum GitPromptFocus {
-    /// Commit the git-repo path as the selected path (same effect as `s`).
-    MountHere,
-    /// Navigate into the repo directory (today's Enter behavior).
-    EnterIn,
-    /// Dismiss the prompt and return to the listing.
-    Cancel,
-}
-
-/// One row in the folder listing.
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub struct FolderEntry {
-    /// Display name, no trailing slash. `".."` for the synthetic parent link.
-    pub name: String,
-    /// Absolute path the row resolves to. For `..` this is the parent dir.
-    pub path: PathBuf,
-    /// True for the synthetic `..` parent-link row.
-    pub is_parent: bool,
-    /// True iff `path` contains a `.git` child (dir or submodule file).
-    pub is_git: bool,
-}
-
-#[derive(Debug)]
-pub struct FileBrowserState {
-    /// $HOME — the browser cannot navigate above this path.
-    pub root: PathBuf,
-    /// Currently-displayed directory.
-    pub cwd: PathBuf,
-    /// Entries loaded from `cwd`, after filtering + sorting.
-    pub entries: Vec<FolderEntry>,
-    /// tui-widget-list selection state. Drives which row is highlighted.
-    pub list_state: ListState,
-    /// Set when the operator presses `s` but the selection is rejected
-    /// (e.g. `$HOME` itself, `~/.jackin/...`). Cleared on the next key.
-    pub rejected_reason: Option<String>,
-    /// Active when the operator has pressed Enter on a git-repo row.
-    pub pending_git_prompt: Option<PathBuf>,
-    /// Origin URL (web form) for the repo referenced by
-    /// `pending_git_prompt`. `None` for non-GitHub remotes or any repo
-    /// whose origin can't be resolved — the overlay then omits the row.
-    pub pending_git_url: Option<String>,
-    /// Which button is highlighted in the git-repo prompt.
-    pub pending_git_focus: GitPromptFocus,
-}
-
-/// Resolve the origin web URL for a git-repo path via `mount_info::inspect`.
-/// Returns `Some` only for GitHub remotes that expose a resolvable web URL.
-fn resolve_git_url(path: &Path) -> Option<String> {
-    match crate::launch::manager::mount_info::inspect(&path.display().to_string()) {
-        crate::launch::manager::mount_info::MountKind::Git { web_url, .. } => web_url,
-        _ => None,
-    }
-}
-
-/// Is `name` one of the top-level noise directories we hide at `$HOME`?
-fn is_excluded(name: &str) -> bool {
-    EXCLUDED.contains(&name)
-}
-
-/// True when `candidate`'s canonicalized path is contained within
-/// `root` (which is already canonicalized when possible). Used to
-/// defeat lexical-prefix spoofs: a symlink under `$HOME` pointing
-/// outside `$HOME` has a lexical path that `Path::starts_with`
-/// accepts, but `canonicalize` resolves the link and exposes the
-/// real target.
-///
-/// Fails closed — any canonicalization error returns `false` rather
-/// than admitting the path. Better to reject a legitimate candidate
-/// (the operator sees an unexplained rejection and can investigate)
-/// than to leak a sandbox-escaping path.
-fn is_within_root(candidate: &Path, root: &Path) -> bool {
-    let Ok(real_candidate) = candidate.canonicalize() else {
-        return false;
-    };
-    real_candidate.starts_with(root)
-}
-
-/// Canonicalize `path`, falling back to `path` itself on error.
-/// Used to normalise `root` once at construction time so later
-/// `is_within_root` comparisons share a common prefix. Production
-/// `$HOME` is almost always canonicalizable; tests under platforms
-/// with weird mount layouts (e.g. macOS `/tmp` → `/private/tmp`)
-/// rely on the fallback to keep behavior sane.
-fn canonicalize_or_self(path: PathBuf) -> PathBuf {
-    path.canonicalize().unwrap_or(path)
-}
-
-/// Read directories under `cwd` and build the entry list. Hidden files
-/// (leading `.`) are excluded; the `..` synthetic parent-link is prepended
-/// iff `cwd != root`; at the root level the `EXCLUDED` list filters out
-/// the macOS-style noise folders. Unreadable directories yield an empty
-/// list — the caller surfaces that as an empty listing, not an error.
-fn load_entries(cwd: &Path, root: &Path) -> Vec<FolderEntry> {
-    let mut out: Vec<FolderEntry> = Vec::new();
-
-    // Synthetic parent link — only shown when there's somewhere to go.
-    if cwd != root
-        && let Some(parent) = cwd.parent()
-        && parent.starts_with(root)
-    {
-        out.push(FolderEntry {
-            name: "..".to_string(),
-            path: parent.to_path_buf(),
-            is_parent: true,
-            is_git: false,
-        });
-    }
-
-    let Ok(read) = std::fs::read_dir(cwd) else {
-        return out;
-    };
-
-    let at_root = cwd == root;
-    let mut dirs: Vec<FolderEntry> = Vec::new();
-    for entry in read.flatten() {
-        let name = entry.file_name().to_string_lossy().to_string();
-        // Skip hidden files (leading dot).
-        if name.starts_with('.') {
-            continue;
-        }
-        // Skip the $HOME-level noise folders.
-        if at_root && is_excluded(&name) {
-            continue;
-        }
-        // Only keep directories.
-        let Ok(ty) = entry.file_type() else {
-            continue;
-        };
-        let path = entry.path();
-        // Follow symlinks to detect dir-symlinks; stat failures skip the entry.
-        let is_dir = ty.is_dir() || (ty.is_symlink() && path.is_dir());
-        if !is_dir {
-            continue;
-        }
-        // Sandbox: a symlinked directory may point outside `root`. Lexical
-        // `starts_with` can't catch that — it sees only the link's path
-        // under root. Canonicalize the entry and drop anything whose real
-        // target escapes. Regular (non-symlink) directories are trusted
-        // because their canonical form necessarily starts with their
-        // parent, which we're already listing because we're inside root.
-        if ty.is_symlink() && !is_within_root(&path, root) {
-            continue;
-        }
-        let is_git = has_git_dir(&path);
-        dirs.push(FolderEntry {
-            name,
-            path,
-            is_parent: false,
-            is_git,
-        });
-    }
-    // Alphabetical, case-insensitive. `..` (if present) was pushed first
-    // and stays at index 0.
-    dirs.sort_by_key(|e| e.name.to_lowercase());
-    out.extend(dirs);
-    out
-}
-
-impl FileBrowserState {
-    /// Build a new browser starting at $HOME, filtered to directories only,
-    /// excluding well-known noisy top-level folders.
-    pub fn new_from_home() -> anyhow::Result<Self> {
-        let home = BaseDirs::new()
-            .map(|b| b.home_dir().to_path_buf())
-            .ok_or_else(|| anyhow::anyhow!("could not resolve $HOME"))?;
-        Ok(Self::new_at(home.clone(), home))
-    }
-
-    /// Build a browser with `root` as the sandbox boundary and `cwd` as
-    /// the initial directory. Primarily for tests — production callers
-    /// should use `new_from_home`.
-    ///
-    /// `root` (and `cwd`, when it was passed as equal to `root`) is
-    /// canonicalized once here so later `is_within_root` checks compare
-    /// apples to apples. If canonicalization fails (exotic mounts,
-    /// missing paths in tests), the uncanonicalized path is used —
-    /// production `$HOME` is always canonicalizable.
-    pub fn new_at(root: PathBuf, cwd: PathBuf) -> Self {
-        let root = canonicalize_or_self(root);
-        // Keep `root == cwd` invariant when the caller intended that —
-        // many tests pass the same path for both, relying on the
-        // "no ..`" affordance at root. Otherwise canonicalize `cwd`
-        // independently so downstream comparisons line up.
-        let cwd = canonicalize_or_self(cwd);
-        let entries = load_entries(&cwd, &root);
-        let mut list_state = ListState::default();
-        if !entries.is_empty() {
-            list_state.select(Some(0));
-        }
-        Self {
-            root,
-            cwd,
-            entries,
-            list_state,
-            rejected_reason: None,
-            pending_git_prompt: None,
-            pending_git_url: None,
-            pending_git_focus: GitPromptFocus::MountHere,
-        }
-    }
-
-    /// Current working directory. Exposed so the create-workspace wizard
-    /// can breadcrumb this across a step-back.
-    pub fn cwd(&self) -> &Path {
-        &self.cwd
-    }
-
-    /// Re-point the browser at `cwd`, clamped to the sandbox root.
-    /// Silently falls back to the root when `cwd` is outside the sandbox
-    /// or not a readable directory.
-    pub fn set_cwd(&mut self, cwd: &Path) {
-        let target = if is_within_root(cwd, &self.root) && cwd.is_dir() {
-            canonicalize_or_self(cwd.to_path_buf())
-        } else {
-            self.root.clone()
-        };
-        self.cwd = target;
-        self.reload();
-    }
-
-    /// Re-read entries from disk and reset the selection to index 0.
-    pub fn reload(&mut self) {
-        self.entries = load_entries(&self.cwd, &self.root);
-        let sel = if self.entries.is_empty() {
-            None
-        } else {
-            Some(0)
-        };
-        self.list_state.select(sel);
-    }
-
-    /// Move selection down one, wrapping at the end.
-    fn select_next(&mut self) {
-        let n = self.entries.len();
-        if n == 0 {
-            return;
-        }
-        let next = self
-            .list_state
-            .selected
-            .map_or(0, |i| if i + 1 >= n { 0 } else { i + 1 });
-        self.list_state.select(Some(next));
-    }
-
-    /// Move selection up one, wrapping at the start.
-    fn select_prev(&mut self) {
-        let n = self.entries.len();
-        if n == 0 {
-            return;
-        }
-        let prev = self
-            .list_state
-            .selected
-            .map_or(0, |i| if i == 0 { n - 1 } else { i - 1 });
-        self.list_state.select(Some(prev));
-    }
-
-    /// The currently-highlighted entry, if any.
-    fn highlighted(&self) -> Option<&FolderEntry> {
-        self.list_state.selected.and_then(|i| self.entries.get(i))
-    }
-
-    /// Navigate up one level (`cwd` → `cwd.parent()`), clamped to `root`.
-    fn navigate_up(&mut self) {
-        if self.cwd == self.root {
-            return;
-        }
-        let Some(parent) = self.cwd.parent() else {
-            return;
-        };
-        if !parent.starts_with(&self.root) {
-            return;
-        }
-        self.cwd = parent.to_path_buf();
-        self.reload();
-    }
-
-    pub fn handle_key(&mut self, key: KeyEvent) -> ModalOutcome<PathBuf> {
-        // Git-repo prompt has its own key map; delegate before clearing
-        // any state the main handler would otherwise reset.
-        if self.pending_git_prompt.is_some() {
-            return self.handle_git_prompt_key(key);
-        }
-
-        // Clear any stale rejection on the next keypress.
-        self.rejected_reason = None;
-
-        match key.code {
-            KeyCode::Up | KeyCode::Char('k' | 'K') => {
-                self.select_prev();
-                ModalOutcome::Continue
-            }
-            KeyCode::Down | KeyCode::Char('j' | 'J') => {
-                self.select_next();
-                ModalOutcome::Continue
-            }
-            KeyCode::Left | KeyCode::Char('h' | 'H') => {
-                self.navigate_up();
-                ModalOutcome::Continue
-            }
-            KeyCode::Enter | KeyCode::Right | KeyCode::Char('l' | 'L') => self.handle_enter(),
-            KeyCode::Char('s' | 'S') => {
-                // Prefer the highlighted entry (so the operator can pick a
-                // sibling without first Entering). Fall back to cwd when
-                // there's no real selection (empty listing, `..` row).
-                let target = self
-                    .highlighted()
-                    .filter(|e| !e.is_parent)
-                    .map_or_else(|| self.cwd.clone(), |e| e.path.clone());
-                self.commit_or_reject(target)
-            }
-            KeyCode::Esc => {
-                // Esc steps back one directory when the operator has
-                // drilled below root — mirroring `h` / `←`. Only cancels
-                // the modal when already at root. `rejected_reason` was
-                // cleared above; preserve that in both branches.
-                if self.cwd == self.root {
-                    ModalOutcome::Cancel
-                } else {
-                    self.navigate_up();
-                    ModalOutcome::Continue
-                }
-            }
-            _ => ModalOutcome::Continue,
-        }
-    }
-
-    /// Shared path for Enter / l / → on a highlighted entry: navigate
-    /// into dirs, open the git-repo prompt on repo rows, up on the `..` row.
-    fn handle_enter(&mut self) -> ModalOutcome<PathBuf> {
-        let Some(entry) = self.highlighted().cloned() else {
-            return ModalOutcome::Continue;
-        };
-        if entry.is_parent {
-            self.navigate_up();
-            return ModalOutcome::Continue;
-        }
-        if entry.is_git {
-            self.pending_git_url = resolve_git_url(&entry.path);
-            self.pending_git_prompt = Some(entry.path);
-            self.pending_git_focus = GitPromptFocus::MountHere;
-            return ModalOutcome::Continue;
-        }
-        // Plain folder — navigate in. Canonicalize so a legitimate
-        // symlinked-dir-inside-root still resolves to its real path;
-        // out-of-root symlinks were already filtered by `load_entries`,
-        // but this belt-and-suspenders guards against races.
-        if is_within_root(&entry.path, &self.root) && entry.path.is_dir() {
-            self.cwd = canonicalize_or_self(entry.path);
-            self.reload();
-        }
-        ModalOutcome::Continue
-    }
-
-    /// Shared commit-or-reject logic used by `s` and the git-repo prompt's
-    /// "Mount this repository" option. Enforces the same sandbox rules.
-    fn commit_or_reject(&mut self, target: PathBuf) -> ModalOutcome<PathBuf> {
-        // Sandbox: belt-and-suspenders check that `target`'s canonical
-        // form is inside `root`. Upstream steps (load_entries, set_cwd,
-        // handle_enter) already filter escaping symlinks, but a TOCTOU
-        // race between listing and commit could still slip one through.
-        if !is_within_root(&target, &self.root) {
-            self.rejected_reason = Some("Cannot commit a path outside $HOME.".into());
-            return ModalOutcome::Continue;
-        }
-        // Reject root itself — user must navigate into a subfolder.
-        if target == self.root {
-            self.rejected_reason =
-                Some("Cannot use $HOME itself — navigate into a subfolder.".into());
-            return ModalOutcome::Continue;
-        }
-        // Reject jackin's own data directory.
-        let jackin_data = self.root.join(".jackin");
-        if target.starts_with(&jackin_data) {
-            self.rejected_reason =
-                Some("Cannot use ~/.jackin/* — those paths are reserved.".into());
-            return ModalOutcome::Continue;
-        }
-        ModalOutcome::Commit(target)
-    }
-
-    /// Clear the git-repo prompt state in one shot — both the pending
-    /// path and the resolved URL.
-    fn dismiss_git_prompt(&mut self) {
-        self.pending_git_prompt = None;
-        self.pending_git_url = None;
-    }
-
-    /// Key handler used while the git-repo prompt is active.
-    fn handle_git_prompt_key(&mut self, key: KeyEvent) -> ModalOutcome<PathBuf> {
-        let Some(path) = self.pending_git_prompt.clone() else {
-            return ModalOutcome::Continue;
-        };
-        match key.code {
-            KeyCode::Char('m' | 'M') => {
-                self.dismiss_git_prompt();
-                self.commit_or_reject(path)
-            }
-            // `p` for "pick a subdirectory" — matches the button label
-            // (renamed from `Enter` to `Pick` in batch 16).
-            KeyCode::Char('p' | 'P') => {
-                self.dismiss_git_prompt();
-                self.set_cwd(&path);
-                ModalOutcome::Continue
-            }
-            // `o` for "open the repo's web URL in the browser" — best-effort;
-            // silent no-op when `pending_git_url` is `None` (non-GitHub origin
-            // or unresolvable remote) or when the launcher fails. The overlay
-            // drops the `· O open` hint segment in the None case so the
-            // keystroke is only advertised when it actually does something.
-            KeyCode::Char('o' | 'O') => {
-                if let Some(url) = self.pending_git_url.as_deref() {
-                    let _ = open::that_detached(url);
-                }
-                ModalOutcome::Continue
-            }
-            KeyCode::Char('c' | 'C') | KeyCode::Esc => {
-                self.dismiss_git_prompt();
-                ModalOutcome::Continue
-            }
-            KeyCode::Enter => {
-                let focus = self.pending_git_focus;
-                self.dismiss_git_prompt();
-                match focus {
-                    GitPromptFocus::MountHere => self.commit_or_reject(path),
-                    GitPromptFocus::EnterIn => {
-                        self.set_cwd(&path);
-                        ModalOutcome::Continue
-                    }
-                    GitPromptFocus::Cancel => ModalOutcome::Continue,
-                }
-            }
-            KeyCode::Tab | KeyCode::Right | KeyCode::Char('l' | 'L') => {
-                self.pending_git_focus = match self.pending_git_focus {
-                    GitPromptFocus::MountHere => GitPromptFocus::EnterIn,
-                    GitPromptFocus::EnterIn => GitPromptFocus::Cancel,
-                    GitPromptFocus::Cancel => GitPromptFocus::MountHere,
-                };
-                ModalOutcome::Continue
-            }
-            KeyCode::BackTab | KeyCode::Left | KeyCode::Char('h' | 'H') => {
-                self.pending_git_focus = match self.pending_git_focus {
-                    GitPromptFocus::MountHere => GitPromptFocus::Cancel,
-                    GitPromptFocus::EnterIn => GitPromptFocus::MountHere,
-                    GitPromptFocus::Cancel => GitPromptFocus::EnterIn,
-                };
-                ModalOutcome::Continue
-            }
-            _ => ModalOutcome::Continue,
-        }
-    }
-}
-
-/// Vertical-layout constraints used by `render` and by the geometry-only
-/// helpers consumed by the mouse-click hit-tester. Keep these in sync.
-fn render_constraints(has_rejection: bool) -> Vec<ratatui::layout::Constraint> {
-    use ratatui::layout::Constraint;
-    if has_rejection {
-        vec![
-            Constraint::Length(1),
-            Constraint::Min(3),
-            Constraint::Length(1),
-        ]
-    } else {
-        vec![Constraint::Min(3), Constraint::Length(1)]
-    }
-}
-
-/// Rect of the listing area inside the modal.
-///
-/// This is the same chunk that `render` passes to `render_listing` and
-/// anchors `render_git_prompt` on. Exposed so a mouse-handler can
-/// recompute the git-prompt overlay geometry without needing `&mut`
-/// access at render time.
-pub fn listing_rect(modal_area: Rect, has_rejection: bool) -> Rect {
-    use ratatui::layout::{Direction, Layout};
-    let constraints = render_constraints(has_rejection);
-    let listing_idx = usize::from(has_rejection);
-    let chunks = Layout::default()
-        .direction(Direction::Vertical)
-        .constraints(constraints)
-        .split(modal_area);
-    chunks[listing_idx]
-}
-
-/// Rect of the git-repo prompt overlay, mirroring the geometry in
-/// `render_git_prompt`. Returns `None` when the overlay would exceed the
-/// listing area.
-pub fn git_prompt_rect(listing: Rect, has_url: bool) -> Option<Rect> {
-    let w = listing.width.saturating_sub(4).min(80);
-    let base_h: u16 = if has_url { 8 } else { 7 };
-    let h = base_h.min(listing.height);
-    if w == 0 || h == 0 {
-        return None;
-    }
-    let x = listing.x + listing.width.saturating_sub(w) / 2;
-    let y = listing.y + listing.height.saturating_sub(h) / 2;
-    Some(Rect {
-        x,
-        y,
-        width: w,
-        height: h,
-    })
-}
-
-/// Rect of the URL row inside the git-prompt overlay, in absolute
-/// screen coordinates. Returns `None` when `has_url` is false — the
-/// URL row isn't rendered then and a click there shouldn't open anything.
-///
-/// Row order inside the overlay's inner (borders stripped) body is
-/// `[prompt][url?][spacer][buttons][spacer][hint]`, all Length(1). So the
-/// URL row sits at `inner.y + 1 = overlay.y + 1 (top border) + 1 = overlay.y + 2`.
-pub fn git_prompt_url_row_rect(modal_area: Rect, has_rejection: bool) -> Option<Rect> {
-    let listing = listing_rect(modal_area, has_rejection);
-    let overlay = git_prompt_rect(listing, true)?;
-    // Need at least borders + prompt + url rows — otherwise the URL row
-    // got clipped by the parent's height.
-    if overlay.height < 3 {
-        return None;
-    }
-    // Inside the block: strip the borders, then take row index 1.
-    let inner_x = overlay.x + 1;
-    let inner_width = overlay.width.saturating_sub(2);
-    let url_y = overlay.y + 2;
-    Some(Rect {
-        x: inner_x,
-        y: url_y,
-        width: inner_width,
-        height: 1,
-    })
-}
-
-impl FileBrowserState {
-    /// Handle a left-click at `(column, row)` in absolute terminal
-    /// coordinates while `modal_area` hosts this browser. Returns `true`
-    /// iff the click hit the git-prompt's URL row AND `pending_git_url`
-    /// is resolved — in which case `open::that_detached` has been fired
-    /// best-effort (errors are swallowed; see the `O` hotkey handler for
-    /// the parallel rationale). A `true` return doesn't dismiss the
-    /// prompt, matching the keyboard keypath.
-    pub fn maybe_open_url_on_click(&self, modal_area: Rect, column: u16, row: u16) -> bool {
-        if self.pending_git_prompt.is_none() {
-            return false;
-        }
-        let Some(url) = self.pending_git_url.as_deref() else {
-            return false;
-        };
-        let has_rejection = self.rejected_reason.is_some();
-        let Some(url_rect) = git_prompt_url_row_rect(modal_area, has_rejection) else {
-            return false;
-        };
-        if column < url_rect.x || column >= url_rect.x + url_rect.width || row != url_rect.y {
-            return false;
-        }
-        let _ = open::that_detached(url);
-        true
-    }
-}
-
-pub fn render(frame: &mut Frame, area: Rect, state: &FileBrowserState) {
-    use ratatui::layout::{Alignment, Direction, Layout};
-
-    frame.render_widget(ratatui::widgets::Clear, area);
-
-    // Layout: [optional rejection banner][listing][nav hint].
-    let has_rejection = state.rejected_reason.is_some();
-    let constraints = render_constraints(has_rejection);
-    let chunks = Layout::default()
-        .direction(Direction::Vertical)
-        .constraints(constraints)
-        .split(area);
-
-    let listing_idx = if has_rejection {
-        let reason = state.rejected_reason.as_ref().unwrap();
-        frame.render_widget(
-            Paragraph::new(Span::styled(
-                format!("\u{2717} {reason}"),
-                Style::default().fg(DANGER_RED).add_modifier(Modifier::BOLD),
-            ))
-            .alignment(Alignment::Center),
-            chunks[0],
-        );
-        1
-    } else {
-        0
-    };
-
-    render_listing(frame, chunks[listing_idx], state);
-    render_footer_legend(frame, chunks[chunks.len() - 1], state);
-
-    // Git-repo prompt overlay — centred inside the listing area so the
-    // listing stays visible as context behind the modal.
-    if state.pending_git_prompt.is_some() {
-        render_git_prompt(frame, chunks[listing_idx], state);
-    }
-}
-
-/// Render the folder listing inside `area` with a phosphor-framed block
-/// and a bold-white cwd title.
-fn render_listing(frame: &mut Frame, area: Rect, state: &FileBrowserState) {
-    let title = Span::styled(
-        format!(
-            " {} ",
-            crate::tui::shorten_home(&state.cwd.display().to_string())
-        ),
-        Style::default().fg(WHITE).add_modifier(Modifier::BOLD),
-    );
-    let block = Block::default()
-        .borders(Borders::ALL)
-        .border_style(Style::default().fg(PHOSPHOR_DARK))
-        .title(title);
-
-    let selected = state.list_state.selected;
-    let highlight_style = Style::default()
-        .bg(PHOSPHOR_GREEN)
-        .fg(Color::Black)
-        .add_modifier(Modifier::BOLD);
-    let base_style = Style::default().fg(WHITE);
-    let git_suffix_style = Style::default()
-        .fg(PHOSPHOR_GREEN)
-        .add_modifier(Modifier::BOLD);
-
-    let lines: Vec<Line> = state
-        .entries
-        .iter()
-        .enumerate()
-        .map(|(i, e)| {
-            let is_sel = Some(i) == selected;
-            let name_slash = if e.is_parent {
-                "../".to_string()
-            } else {
-                format!("{}/", e.name)
-            };
-            if is_sel {
-                // Highlight row: single span covering name + optional git suffix.
-                let mut text = format!("  {name_slash}");
-                if e.is_git {
-                    text.push_str(" (git)");
-                }
-                Line::from(Span::styled(text, highlight_style))
-            } else if e.is_git {
-                Line::from(vec![
-                    Span::styled(format!("  {name_slash}"), base_style),
-                    Span::styled(" (git)", git_suffix_style),
-                ])
-            } else {
-                Line::from(Span::styled(format!("  {name_slash}"), base_style))
-            }
-        })
-        .collect();
-
-    let paragraph = Paragraph::new(lines).block(block);
-    frame.render_widget(paragraph, area);
-}
-
-/// Render the bottom footer legend. Swaps the usual nav+`s` legend for a
-/// prompt-focused legend when the git-repo confirm overlay is active.
-fn render_footer_legend(frame: &mut Frame, area: Rect, state: &FileBrowserState) {
-    use ratatui::layout::Alignment;
-    let key = Style::default().fg(WHITE).add_modifier(Modifier::BOLD);
-    let text = Style::default().fg(PHOSPHOR_GREEN);
-    let sep = Style::default().fg(PHOSPHOR_DARK);
-    let line = if state.pending_git_prompt.is_some() {
-        Line::from(vec![
-            Span::styled("Enter", key),
-            Span::styled(" confirm", text),
-            Span::styled(" \u{b7} ", sep),
-            Span::styled("Esc", key),
-            Span::styled(" cancel", text),
-        ])
-    } else {
-        Line::from(vec![
-            Span::styled("\u{2191}\u{2193}", key),
-            Span::styled(" navigate", text),
-            Span::styled(" \u{b7} ", sep),
-            Span::styled("Enter", key),
-            Span::styled(" open", text),
-            Span::styled(" \u{b7} ", sep),
-            Span::styled("H/\u{2190}", key),
-            Span::styled(" up", text),
-            Span::raw("   "),
-            Span::styled("S", key),
-            Span::styled(" select", text),
-            Span::raw("   "),
-            Span::styled("Esc", key),
-            Span::styled(" up/cancel", text),
-        ])
-    };
-    frame.render_widget(Paragraph::new(line).alignment(Alignment::Center), area);
-}
-
-/// Build the three focus-styled button spans for the git-repo prompt.
-/// Focused choice highlights on white; unfocused stays flush with the
-/// modal background so only the focused choice pops (canonical template).
-fn git_prompt_buttons(focus: GitPromptFocus) -> Line<'static> {
-    let focused = Style::default()
-        .bg(WHITE)
-        .fg(Color::Black)
-        .add_modifier(Modifier::BOLD);
-    let unfocused = Style::default()
-        .fg(PHOSPHOR_GREEN)
-        .add_modifier(Modifier::BOLD);
-    let btn = |target: GitPromptFocus, label: &'static str| -> Span<'static> {
-        let style = if focus == target { focused } else { unfocused };
-        Span::styled(format!("  {label}  "), style)
-    };
-    Line::from(vec![
-        btn(GitPromptFocus::MountHere, "Mount this repository"),
-        Span::raw("    "),
-        btn(GitPromptFocus::EnterIn, "Pick a subdirectory"),
-        Span::raw("    "),
-        btn(GitPromptFocus::Cancel, "Cancel"),
-    ])
-}
-
-/// Build the canonical hint footer line for the git-repo prompt.
-///
-/// When `has_url` is true:
-/// `M mount · P pick · O open · C/Esc cancel`.
-/// When `has_url` is false, the `· O open` segment is dropped so the
-/// hint doesn't advertise a disabled action:
-/// `M mount · P pick · C/Esc cancel`.
-fn git_prompt_hint(has_url: bool) -> Line<'static> {
-    let key_style = Style::default().fg(WHITE).add_modifier(Modifier::BOLD);
-    let text_style = Style::default().fg(PHOSPHOR_GREEN);
-    let sep_style = Style::default().fg(PHOSPHOR_DARK);
-    let mut spans: Vec<Span<'static>> = vec![
-        Span::styled("M", key_style),
-        Span::styled(" mount", text_style),
-        Span::styled(" \u{b7} ", sep_style),
-        Span::styled("P", key_style),
-        Span::styled(" pick", text_style),
-    ];
-    if has_url {
-        spans.extend([
-            Span::styled(" \u{b7} ", sep_style),
-            Span::styled("O", key_style),
-            Span::styled(" open", text_style),
-        ]);
-    }
-    spans.extend([
-        Span::styled(" \u{b7} ", sep_style),
-        Span::styled("C/Esc", key_style),
-        Span::styled(" cancel", text_style),
-    ]);
-    Line::from(spans)
-}
-
-/// Overlay renderer for the in-browser "Git repository detected" prompt.
-fn render_git_prompt(frame: &mut Frame, parent: Rect, state: &FileBrowserState) {
-    use ratatui::layout::{Alignment, Constraint, Direction, Layout};
-
-    // Add a row when we have an origin URL to show under the title.
-    let has_url = state.pending_git_url.is_some();
-    // Overlay widens to 80 cols so the three-button row and the canonical
-    // hint line both fit on one line without wrapping.
-    let w = parent.width.saturating_sub(4).min(80);
-    let base_h: u16 = if has_url { 8 } else { 7 };
-    let h = base_h.min(parent.height);
-    let x = parent.x + parent.width.saturating_sub(w) / 2;
-    let y = parent.y + parent.height.saturating_sub(h) / 2;
-    let area = Rect {
-        x,
-        y,
-        width: w,
-        height: h,
-    };
-
-    let block = Block::default()
-        .borders(Borders::ALL)
-        .border_style(Style::default().fg(PHOSPHOR_DARK))
-        .title(Span::styled(
-            " Git repository detected ",
-            Style::default().fg(WHITE).add_modifier(Modifier::BOLD),
-        ));
-    let inner = block.inner(area);
-    frame.render_widget(ratatui::widgets::Clear, area);
-    frame.render_widget(block, area);
-
-    // Row constraints: [prompt][url?][spacer][buttons][spacer][hint].
-    let row_count = if has_url { 6 } else { 5 };
-    let constraints: Vec<Constraint> = (0..row_count).map(|_| Constraint::Length(1)).collect();
-    let rows = Layout::default()
-        .direction(Direction::Vertical)
-        .constraints(constraints)
-        .split(inner);
-
-    frame.render_widget(
-        Paragraph::new(Span::styled(
-            "What would you like to do?",
-            Style::default().fg(WHITE).add_modifier(Modifier::BOLD),
-        ))
-        .alignment(Alignment::Center),
-        rows[0],
-    );
-
-    let (buttons_idx, hint_idx) = if has_url {
-        let url = state.pending_git_url.as_deref().unwrap_or_default();
-        frame.render_widget(
-            Paragraph::new(Span::styled(
-                url.to_string(),
-                Style::default()
-                    .fg(PHOSPHOR_DIM)
-                    .add_modifier(Modifier::ITALIC),
-            ))
-            .alignment(Alignment::Center),
-            rows[1],
-        );
-        (3, 5)
-    } else {
-        (2, 4)
-    };
-
-    frame.render_widget(
-        Paragraph::new(git_prompt_buttons(state.pending_git_focus)).alignment(Alignment::Center),
-        rows[buttons_idx],
-    );
-    frame.render_widget(
-        Paragraph::new(git_prompt_hint(has_url)).alignment(Alignment::Center),
-        rows[hint_idx],
-    );
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use crossterm::event::{KeyEvent, KeyEventKind, KeyEventState, KeyModifiers};
-    use tempfile::tempdir;
-
-    fn key(code: KeyCode) -> KeyEvent {
-        KeyEvent {
-            code,
-            modifiers: KeyModifiers::NONE,
-            kind: KeyEventKind::Press,
-            state: KeyEventState::NONE,
-        }
-    }
-
-    fn make_state_at(path: PathBuf) -> FileBrowserState {
-        FileBrowserState::new_at(path.clone(), path)
-    }
-
-    fn state_rooted_at(root: PathBuf, cwd: PathBuf) -> FileBrowserState {
-        FileBrowserState::new_at(root, cwd)
-    }
-
-    // ── Filtering + directory-only listing ────────────────────────────
-
-    #[test]
-    fn filter_excludes_files() {
-        let tmp = tempdir().unwrap();
-        std::fs::create_dir(tmp.path().join("folder")).unwrap();
-        std::fs::write(tmp.path().join("file.txt"), b"x").unwrap();
-
-        let state = make_state_at(tmp.path().to_path_buf());
-        let names: Vec<&str> = state.entries.iter().map(|e| e.name.as_str()).collect();
-        assert!(names.contains(&"folder"), "folder missing: {names:?}");
-        assert!(
-            !names.contains(&"file.txt"),
-            "file should be filtered out: {names:?}"
-        );
-    }
-
-    #[test]
-    fn hidden_files_are_excluded() {
-        let tmp = tempdir().unwrap();
-        std::fs::create_dir(tmp.path().join("visible")).unwrap();
-        std::fs::create_dir(tmp.path().join(".hidden")).unwrap();
-
-        let state = make_state_at(tmp.path().to_path_buf());
-        let names: Vec<&str> = state.entries.iter().map(|e| e.name.as_str()).collect();
-        assert!(names.contains(&"visible"));
-        assert!(!names.contains(&".hidden"));
-    }
-
-    #[test]
-    fn excluded_names_filtered_at_root() {
-        let tmp = tempdir().unwrap();
-        for name in EXCLUDED {
-            std::fs::create_dir(tmp.path().join(name)).unwrap();
-        }
-        std::fs::create_dir(tmp.path().join("Projects")).unwrap();
-
-        let state = make_state_at(tmp.path().to_path_buf());
-        let names: Vec<&str> = state.entries.iter().map(|e| e.name.as_str()).collect();
-        for name in EXCLUDED {
-            assert!(!names.contains(name), "excluded `{name}` slipped through");
-        }
-        assert!(names.contains(&"Projects"));
-    }
-
-    #[test]
-    fn excluded_names_visible_below_root() {
-        // EXCLUDED only applies at the sandbox root; a folder named
-        // "Library" one level below should still be visible.
-        let tmp = tempdir().unwrap();
-        let sub = tmp.path().join("sub");
-        std::fs::create_dir_all(sub.join("Library")).unwrap();
-
-        let state = state_rooted_at(tmp.path().to_path_buf(), sub);
-        let names: Vec<&str> = state.entries.iter().map(|e| e.name.as_str()).collect();
-        assert!(names.contains(&"Library"));
-    }
-
-    #[test]
-    fn parent_link_absent_at_root() {
-        let tmp = tempdir().unwrap();
-        std::fs::create_dir(tmp.path().join("a")).unwrap();
-        let state = make_state_at(tmp.path().to_path_buf());
-        assert!(
-            state.entries.iter().all(|e| !e.is_parent),
-            "`..` must not appear at root: {:?}",
-            state.entries
-        );
-    }
-
-    #[test]
-    fn parent_link_present_below_root() {
-        let tmp = tempdir().unwrap();
-        let sub = tmp.path().join("sub");
-        std::fs::create_dir(&sub).unwrap();
-        let state = state_rooted_at(tmp.path().to_path_buf(), sub);
-        assert!(state.entries.first().is_some_and(|e| e.is_parent));
-    }
-
-    // ── Git-repo detection ────────────────────────────────────────────
-
-    #[test]
-    fn git_repo_entries_have_is_git_true() {
-        let tmp = tempdir().unwrap();
-        let repo = tmp.path().join("repo");
-        std::fs::create_dir_all(repo.join(".git")).unwrap();
-
-        let state = make_state_at(tmp.path().to_path_buf());
-        let entry = state
-            .entries
-            .iter()
-            .find(|e| e.name == "repo")
-            .expect("repo row must exist");
-        assert!(entry.is_git, "repo row must be flagged as git");
-    }
-
-    #[test]
-    fn non_git_folders_have_is_git_false() {
-        let tmp = tempdir().unwrap();
-        let plain = tmp.path().join("plain");
-        std::fs::create_dir(&plain).unwrap();
-
-        let state = make_state_at(tmp.path().to_path_buf());
-        let entry = state
-            .entries
-            .iter()
-            .find(|e| e.name == "plain")
-            .expect("plain row must exist");
-        assert!(!entry.is_git);
-    }
-
-    #[test]
-    fn submodule_gitfile_counts_as_git() {
-        let tmp = tempdir().unwrap();
-        let sub = tmp.path().join("submodule");
-        std::fs::create_dir(&sub).unwrap();
-        std::fs::write(sub.join(".git"), "gitdir: ../.git/modules/submodule\n").unwrap();
-
-        let state = make_state_at(tmp.path().to_path_buf());
-        let entry = state
-            .entries
-            .iter()
-            .find(|e| e.name == "submodule")
-            .expect("submodule row must exist");
-        assert!(entry.is_git);
-    }
-
-    // ── Render: ensure the ` (git)` suffix actually appears ───────────
-
-    #[test]
-    fn git_entries_render_with_git_suffix() {
-        use ratatui::{Terminal, backend::TestBackend};
-
-        let tmp = tempdir().unwrap();
-        let repo = tmp.path().join("repo");
-        std::fs::create_dir_all(repo.join(".git")).unwrap();
-        std::fs::create_dir(tmp.path().join("plain")).unwrap();
-
-        // Use a state where the selection is NOT on the git row, so the
-        // suffix renders as a separate span rather than getting absorbed
-        // into the highlight style.
-        let mut state = make_state_at(tmp.path().to_path_buf());
-        // Sort order is alphabetical lowercase: plain < repo. Select plain
-        // (index 0) so repo's ` (git)` suffix renders unhighlighted.
-        state.list_state.select(Some(0));
-
-        let backend = TestBackend::new(40, 10);
-        let mut terminal = Terminal::new(backend).unwrap();
-        terminal
-            .draw(|frame| {
-                render(frame, frame.area(), &state);
-            })
-            .unwrap();
-
-        let buffer = terminal.backend().buffer();
-        let dump = buffer
-            .content()
-            .iter()
-            .map(|cell| cell.symbol())
-            .collect::<String>();
-        assert!(dump.contains("repo/"), "repo row should render: {dump:?}");
-        assert!(
-            dump.contains("(git)"),
-            "git suffix should render on the repo row: {dump:?}"
-        );
-        assert!(dump.contains("plain/"));
-    }
-
-    // ── `s` behaviour ─────────────────────────────────────────────────
-
-    #[test]
-    fn s_commits_highlighted_entry() {
-        let tmp = tempdir().unwrap();
-        let parent = tmp.path().join("parent");
-        let child = parent.join("child");
-        std::fs::create_dir_all(&child).unwrap();
-
-        // root = tmp so that neither parent nor child trip the $HOME guard.
-        let mut state = state_rooted_at(tmp.path().to_path_buf(), parent.clone());
-        // Highlighted entry at index 0 is `..`; advance to `child`.
-        state.handle_key(key(KeyCode::Down));
-
-        let outcome = state.handle_key(key(KeyCode::Char('s')));
-        match outcome {
-            ModalOutcome::Commit(path) => {
-                assert_eq!(path.canonicalize().unwrap(), child.canonicalize().unwrap(),);
-            }
-            other => panic!("expected Commit, got {other:?}"),
-        }
-    }
-
-    #[test]
-    fn s_falls_back_to_cwd_when_directory_is_empty() {
-        let tmp = tempdir().unwrap();
-        let empty = tmp.path().join("empty");
-        std::fs::create_dir(&empty).unwrap();
-
-        let mut state = state_rooted_at(tmp.path().to_path_buf(), empty.clone());
-        // Empty except for `..` — `s` should commit cwd, not `..`.
-        let outcome = state.handle_key(key(KeyCode::Char('s')));
-        match outcome {
-            ModalOutcome::Commit(path) => {
-                assert_eq!(path.canonicalize().unwrap(), empty.canonicalize().unwrap(),);
-            }
-            other => panic!("expected Commit, got {other:?}"),
-        }
-    }
-
-    #[test]
-    fn s_rejects_root_itself() {
-        let tmp = tempdir().unwrap();
-        let mut state = make_state_at(tmp.path().to_path_buf());
-        let outcome = state.handle_key(key(KeyCode::Char('s')));
-        assert!(matches!(outcome, ModalOutcome::Continue));
-        assert!(state.rejected_reason.is_some());
-    }
-
-    #[test]
-    fn s_rejects_jackin_data_dir() {
-        let tmp = tempdir().unwrap();
-        let jackin = tmp.path().join(".jackin").join("workspaces");
-        std::fs::create_dir_all(&jackin).unwrap();
-
-        let mut state = state_rooted_at(tmp.path().to_path_buf(), jackin);
-        let outcome = state.handle_key(key(KeyCode::Char('s')));
-        assert!(matches!(outcome, ModalOutcome::Continue));
-        assert!(state.rejected_reason.is_some());
-    }
-
-    #[test]
-    fn rejection_cleared_on_next_keypress() {
-        let tmp = tempdir().unwrap();
-        let mut state = make_state_at(tmp.path().to_path_buf());
-        state.handle_key(key(KeyCode::Char('s')));
-        assert!(state.rejected_reason.is_some());
-        state.handle_key(key(KeyCode::Char('j')));
-        assert!(state.rejected_reason.is_none());
-    }
-
-    // ── Esc step-back navigation ──────────────────────────────────────
-
-    #[test]
-    fn esc_at_root_cancels_modal() {
-        let tmp = tempdir().unwrap();
-        let mut state = make_state_at(tmp.path().to_path_buf());
-        let outcome = state.handle_key(key(KeyCode::Esc));
-        assert!(matches!(outcome, ModalOutcome::Cancel));
-    }
-
-    #[test]
-    fn esc_inside_subfolder_navigates_up() {
-        let tmp = tempdir().unwrap();
-        let sub = tmp.path().join("sub");
-        std::fs::create_dir(&sub).unwrap();
-
-        let mut state = state_rooted_at(tmp.path().to_path_buf(), sub.clone());
-        let outcome = state.handle_key(key(KeyCode::Esc));
-        assert!(matches!(outcome, ModalOutcome::Continue));
-        assert_eq!(
-            state.cwd.canonicalize().unwrap(),
-            tmp.path().canonicalize().unwrap(),
-        );
-    }
-
-    #[test]
-    fn esc_deep_navigates_up_one_level() {
-        let tmp = tempdir().unwrap();
-        let l1 = tmp.path().join("a");
-        let l2 = l1.join("b");
-        let l3 = l2.join("c");
-        std::fs::create_dir_all(&l3).unwrap();
-
-        let mut state = state_rooted_at(tmp.path().to_path_buf(), l3);
-        state.handle_key(key(KeyCode::Esc));
-        assert_eq!(
-            state.cwd.canonicalize().unwrap(),
-            l2.canonicalize().unwrap(),
-        );
-    }
-
-    #[test]
-    fn esc_clears_rejected_reason() {
-        let tmp = tempdir().unwrap();
-        let mut state = make_state_at(tmp.path().to_path_buf());
-        state.rejected_reason = Some("stale reason".into());
-        let outcome = state.handle_key(key(KeyCode::Esc));
-        assert!(matches!(outcome, ModalOutcome::Cancel));
-        assert!(state.rejected_reason.is_none());
-    }
-
-    // ── h / l navigation ──────────────────────────────────────────────
-
-    #[test]
-    fn h_navigates_up() {
-        let tmp = tempdir().unwrap();
-        let sub = tmp.path().join("sub");
-        std::fs::create_dir(&sub).unwrap();
-        let mut state = state_rooted_at(tmp.path().to_path_buf(), sub);
-        state.handle_key(key(KeyCode::Char('h')));
-        assert_eq!(
-            state.cwd.canonicalize().unwrap(),
-            tmp.path().canonicalize().unwrap(),
-        );
-    }
-
-    #[test]
-    fn h_at_root_is_noop() {
-        let tmp = tempdir().unwrap();
-        let mut state = make_state_at(tmp.path().to_path_buf());
-        state.handle_key(key(KeyCode::Char('h')));
-        assert_eq!(state.cwd, tmp.path());
-    }
-
-    #[test]
-    fn l_navigates_into_highlighted_dir() {
-        let tmp = tempdir().unwrap();
-        let child = tmp.path().join("child");
-        std::fs::create_dir(&child).unwrap();
-        let mut state = make_state_at(tmp.path().to_path_buf());
-        // No `..` at root — index 0 is `child`.
-        state.handle_key(key(KeyCode::Char('l')));
-        assert_eq!(
-            state.cwd.canonicalize().unwrap(),
-            child.canonicalize().unwrap(),
-        );
-    }
-
-    // ── Git-repo prompt ───────────────────────────────────────────────
-
-    #[test]
-    fn enter_on_git_repo_opens_prompt() {
-        let tmp = tempdir().unwrap();
-        let parent = tmp.path().join("parent");
-        let repo = parent.join("repo");
-        std::fs::create_dir_all(repo.join(".git")).unwrap();
-
-        let mut state = state_rooted_at(tmp.path().to_path_buf(), parent);
-        // Index 0 is `..`; advance to `repo`.
-        state.handle_key(key(KeyCode::Down));
-        let outcome = state.handle_key(key(KeyCode::Enter));
-        assert!(matches!(outcome, ModalOutcome::Continue));
-        assert!(state.pending_git_prompt.is_some());
-        assert_eq!(state.pending_git_focus, GitPromptFocus::MountHere);
-    }
-
-    /// Write a minimal `.git/HEAD` + `.git/config` with the given origin.
-    fn seed_git_repo_with_origin(repo: &Path, remote: &str) {
-        let git = repo.join(".git");
-        std::fs::create_dir_all(&git).unwrap();
-        std::fs::write(git.join("HEAD"), "ref: refs/heads/main\n").unwrap();
-        std::fs::write(
-            git.join("config"),
-            format!("[remote \"origin\"]\n\turl = {remote}\n"),
-        )
-        .unwrap();
-    }
-
-    #[test]
-    fn enter_on_git_repo_with_origin_sets_url() {
-        let tmp = tempdir().unwrap();
-        let parent = tmp.path().join("parent");
-        let repo = parent.join("repo");
-        std::fs::create_dir_all(&repo).unwrap();
-        seed_git_repo_with_origin(&repo, "git@github.com:jackin-project/jackin.git");
-
-        let mut state = state_rooted_at(tmp.path().to_path_buf(), parent);
-        state.handle_key(key(KeyCode::Down));
-        state.handle_key(key(KeyCode::Enter));
-        assert!(state.pending_git_prompt.is_some());
-        let url = state
-            .pending_git_url
-            .as_deref()
-            .expect("GitHub origin must resolve");
-        assert!(url.contains("github.com/jackin-project/jackin"));
-    }
-
-    #[test]
-    fn enter_on_git_repo_without_origin_leaves_url_none() {
-        let tmp = tempdir().unwrap();
-        let parent = tmp.path().join("parent");
-        let repo = parent.join("repo");
-        std::fs::create_dir_all(repo.join(".git")).unwrap();
-
-        let mut state = state_rooted_at(tmp.path().to_path_buf(), parent);
-        state.handle_key(key(KeyCode::Down));
-        state.handle_key(key(KeyCode::Enter));
-        assert!(state.pending_git_prompt.is_some());
-        assert!(state.pending_git_url.is_none());
-    }
-
-    #[test]
-    fn mount_here_commits_git_path() {
-        let tmp = tempdir().unwrap();
-        let parent = tmp.path().join("parent");
-        let repo = parent.join("repo");
-        std::fs::create_dir_all(repo.join(".git")).unwrap();
-
-        let mut state = state_rooted_at(tmp.path().to_path_buf(), parent);
-        state.handle_key(key(KeyCode::Down));
-        state.handle_key(key(KeyCode::Enter));
-        assert_eq!(state.pending_git_focus, GitPromptFocus::MountHere);
-        let outcome = state.handle_key(key(KeyCode::Enter));
-        match outcome {
-            ModalOutcome::Commit(p) => {
-                assert_eq!(p.canonicalize().unwrap(), repo.canonicalize().unwrap(),);
-            }
-            other => panic!("expected Commit, got {other:?}"),
-        }
-        assert!(state.pending_git_prompt.is_none());
-    }
-
-    #[test]
-    fn enter_in_navigates_into_subdir() {
-        let tmp = tempdir().unwrap();
-        let parent = tmp.path().join("parent");
-        let repo = parent.join("repo");
-        std::fs::create_dir_all(repo.join(".git")).unwrap();
-        std::fs::create_dir(repo.join("sub")).unwrap();
-
-        let mut state = state_rooted_at(tmp.path().to_path_buf(), parent);
-        state.handle_key(key(KeyCode::Down));
-        state.handle_key(key(KeyCode::Enter)); // open prompt
-        state.handle_key(key(KeyCode::Tab)); // MountHere -> EnterIn
-        assert_eq!(state.pending_git_focus, GitPromptFocus::EnterIn);
-
-        let outcome = state.handle_key(key(KeyCode::Enter));
-        assert!(matches!(outcome, ModalOutcome::Continue));
-        assert!(state.pending_git_prompt.is_none());
-        assert_eq!(
-            state.cwd.canonicalize().unwrap(),
-            repo.canonicalize().unwrap(),
-        );
-    }
-
-    #[test]
-    fn cancel_dismisses_prompt_via_focus() {
-        let tmp = tempdir().unwrap();
-        let parent = tmp.path().join("parent");
-        let repo = parent.join("repo");
-        std::fs::create_dir_all(repo.join(".git")).unwrap();
-
-        let mut state = state_rooted_at(tmp.path().to_path_buf(), parent.clone());
-        state.handle_key(key(KeyCode::Down));
-        state.handle_key(key(KeyCode::Enter));
-        state.handle_key(key(KeyCode::Tab));
-        state.handle_key(key(KeyCode::Tab));
-        assert_eq!(state.pending_git_focus, GitPromptFocus::Cancel);
-
-        let outcome = state.handle_key(key(KeyCode::Enter));
-        assert!(matches!(outcome, ModalOutcome::Continue));
-        assert!(state.pending_git_prompt.is_none());
-        assert_eq!(
-            state.cwd.canonicalize().unwrap(),
-            parent.canonicalize().unwrap(),
-        );
-    }
-
-    #[test]
-    fn esc_dismisses_prompt_without_cancelling_browser() {
-        let tmp = tempdir().unwrap();
-        let parent = tmp.path().join("parent");
-        let repo = parent.join("repo");
-        std::fs::create_dir_all(repo.join(".git")).unwrap();
-
-        let mut state = state_rooted_at(tmp.path().to_path_buf(), parent);
-        state.handle_key(key(KeyCode::Down));
-        state.handle_key(key(KeyCode::Enter));
-        assert!(state.pending_git_prompt.is_some());
-        let outcome = state.handle_key(key(KeyCode::Esc));
-        assert!(matches!(outcome, ModalOutcome::Continue));
-        assert!(state.pending_git_prompt.is_none());
-    }
-
-    #[test]
-    fn enter_on_plain_folder_still_navigates() {
-        let tmp = tempdir().unwrap();
-        let parent = tmp.path().join("parent");
-        let plain = parent.join("plain");
-        std::fs::create_dir_all(&plain).unwrap();
-
-        let mut state = state_rooted_at(tmp.path().to_path_buf(), parent);
-        state.handle_key(key(KeyCode::Down));
-        state.handle_key(key(KeyCode::Enter));
-        assert!(state.pending_git_prompt.is_none());
-        assert_eq!(
-            state.cwd.canonicalize().unwrap(),
-            plain.canonicalize().unwrap(),
-        );
-    }
-
-    #[test]
-    fn m_shortcut_commits_repo_from_prompt() {
-        let tmp = tempdir().unwrap();
-        let parent = tmp.path().join("parent");
-        let repo = parent.join("repo");
-        std::fs::create_dir_all(repo.join(".git")).unwrap();
-
-        let mut state = state_rooted_at(tmp.path().to_path_buf(), parent);
-        state.handle_key(key(KeyCode::Down));
-        state.handle_key(key(KeyCode::Enter));
-        state.handle_key(key(KeyCode::Tab));
-        let outcome = state.handle_key(key(KeyCode::Char('m')));
-        match outcome {
-            ModalOutcome::Commit(p) => {
-                assert_eq!(p.canonicalize().unwrap(), repo.canonicalize().unwrap(),);
-            }
-            other => panic!("expected Commit, got {other:?}"),
-        }
-    }
-
-    // ── O hotkey (open URL in browser) ────────────────────────────────
-
-    /// With `pending_git_url == None`, `O` must be a silent no-op: the
-    /// prompt stays open, focus is unchanged, and no commit/cancel fires.
-    /// We can't assert that `open::that_detached` *didn't* run (it doesn't
-    /// run when the URL is None — that's the code path we're testing),
-    /// but we can pin the observable state.
-    #[test]
-    fn o_shortcut_without_url_is_silent_noop() {
-        let tmp = tempdir().unwrap();
-        let parent = tmp.path().join("parent");
-        let repo = parent.join("repo");
-        std::fs::create_dir_all(repo.join(".git")).unwrap();
-
-        let mut state = state_rooted_at(tmp.path().to_path_buf(), parent);
-        state.handle_key(key(KeyCode::Down));
-        state.handle_key(key(KeyCode::Enter));
-        assert!(state.pending_git_prompt.is_some());
-        assert!(state.pending_git_url.is_none());
-        let focus_before = state.pending_git_focus;
-
-        let outcome = state.handle_key(key(KeyCode::Char('o')));
-        assert!(matches!(outcome, ModalOutcome::Continue));
-        // Prompt still open, focus unchanged.
-        assert!(state.pending_git_prompt.is_some());
-        assert_eq!(state.pending_git_focus, focus_before);
-    }
-
-    /// With `pending_git_url == Some(url)`, `O` still returns Continue
-    /// and keeps the prompt open (open-in-browser is best-effort and
-    /// doesn't dismiss). Hidden file:// URL so `open::that_detached` is
-    /// a silent no-op in CI even when a GUI isn't available.
-    #[test]
-    fn o_shortcut_with_url_returns_continue_and_keeps_prompt_open() {
-        let tmp = tempdir().unwrap();
-        let parent = tmp.path().join("parent");
-        let repo = parent.join("repo");
-        std::fs::create_dir_all(repo.join(".git")).unwrap();
-
-        let mut state = state_rooted_at(tmp.path().to_path_buf(), parent);
-        state.handle_key(key(KeyCode::Down));
-        state.handle_key(key(KeyCode::Enter));
-        // Force a URL into state for the test; the real handler would have
-        // populated this via `resolve_git_url` when origin is a GitHub URL.
-        state.pending_git_url = Some("file:///tmp/definitely-not-real".to_string());
-
-        let outcome = state.handle_key(key(KeyCode::Char('O')));
-        assert!(matches!(outcome, ModalOutcome::Continue));
-        assert!(state.pending_git_prompt.is_some());
-        // URL stays on state — O doesn't dismiss the prompt.
-        assert_eq!(
-            state.pending_git_url.as_deref(),
-            Some("file:///tmp/definitely-not-real"),
-        );
-    }
-
-    // ── Conditional hint segment ──────────────────────────────────────
-
-    /// The `O open` hint segment is only rendered when a URL is resolved.
-    /// With `has_url == false` the hint must not advertise `O open`.
-    #[test]
-    fn git_prompt_hint_omits_open_segment_when_url_is_none() {
-        let line = git_prompt_hint(false);
-        let rendered: String = line.spans.iter().map(|s| s.content.as_ref()).collect();
-        assert!(
-            !rendered.contains('O'),
-            "hint should not mention O when no URL: {rendered:?}"
-        );
-        assert!(
-            !rendered.contains("open"),
-            "hint should not mention 'open' when no URL: {rendered:?}"
-        );
-        assert!(rendered.contains('M'));
-        assert!(rendered.contains('P'));
-        assert!(rendered.contains("C/Esc"));
-    }
-
-    #[test]
-    fn git_prompt_hint_includes_open_segment_when_url_is_present() {
-        let line = git_prompt_hint(true);
-        let rendered: String = line.spans.iter().map(|s| s.content.as_ref()).collect();
-        assert!(
-            rendered.contains('O'),
-            "hint should mention O when URL resolved: {rendered:?}"
-        );
-        assert!(
-            rendered.contains("open"),
-            "hint should mention 'open' when URL resolved: {rendered:?}"
-        );
-        // Still preserves the other segments + trailing cancel.
-        assert!(rendered.contains('M'));
-        assert!(rendered.contains('P'));
-        assert!(rendered.contains("C/Esc"));
-    }
-
-    // ── Mouse-click hit-testing on the URL row ────────────────────────
-
-    /// Reference geometry: a 70%-wide, 22-row `modal_area` at term 120x40
-    /// gives `modal_area = Rect { x: 18, y: 9, width: 84, height: 22 }`.
-    /// The listing chunk (no rejection) = `modal_area` minus a 1-row footer
-    /// = `Rect { x: 18, y: 9, width: 84, height: 21 }`. Git-prompt overlay
-    /// width = `min(84-4, 80) = 80`, height = 8 (`has_url = true`), centered
-    /// inside listing → rect x ≈ 20, y = 9 + (21-8)/2 = 15. URL row sits
-    /// at y + 2 = 17.
-    fn manufactured_modal_area() -> Rect {
-        // Mirrors `file_browser_modal_rect` for a term of 120x40:
-        //   w = 120 * 70 / 100 = 84; h = 22.
-        //   x = 0 + (120 - 84)/2 = 18; y = 0 + (40 - 22)/2 = 9.
-        Rect {
-            x: 18,
-            y: 9,
-            width: 84,
-            height: 22,
-        }
-    }
-
-    #[test]
-    fn url_row_rect_none_when_no_url_flag() {
-        // The public helper is parameterised on has_rejection; it always
-        // assumes the git-prompt would render with a URL. This test pins
-        // the returned rect when the overlay would have a URL row.
-        let rect = git_prompt_url_row_rect(manufactured_modal_area(), false);
-        assert!(rect.is_some(), "URL row should resolve for a valid modal");
-    }
-
-    #[test]
-    fn click_on_url_row_without_url_returns_false() {
-        let tmp = tempdir().unwrap();
-        let parent = tmp.path().join("parent");
-        let repo = parent.join("repo");
-        std::fs::create_dir_all(repo.join(".git")).unwrap();
-
-        let mut state = state_rooted_at(tmp.path().to_path_buf(), parent);
-        state.handle_key(key(KeyCode::Down));
-        state.handle_key(key(KeyCode::Enter));
-        assert!(state.pending_git_prompt.is_some());
-        assert!(state.pending_git_url.is_none());
-
-        // Click at the URL row's rough centre — still false because no URL.
-        let modal = manufactured_modal_area();
-        let url_rect = git_prompt_url_row_rect(modal, false).unwrap();
-        let opened =
-            state.maybe_open_url_on_click(modal, url_rect.x + url_rect.width / 2, url_rect.y);
-        assert!(!opened, "click should not open when no URL is resolved");
-    }
-
-    #[test]
-    fn click_outside_url_row_returns_false_even_with_url() {
-        let tmp = tempdir().unwrap();
-        let parent = tmp.path().join("parent");
-        let repo = parent.join("repo");
-        std::fs::create_dir_all(repo.join(".git")).unwrap();
-
-        let mut state = state_rooted_at(tmp.path().to_path_buf(), parent);
-        state.handle_key(key(KeyCode::Down));
-        state.handle_key(key(KeyCode::Enter));
-        state.pending_git_url = Some("file:///tmp/definitely-not-real".to_string());
-
-        let modal = manufactured_modal_area();
-        let url_rect = git_prompt_url_row_rect(modal, false).unwrap();
-        // One row below the URL row — outside.
-        let opened =
-            state.maybe_open_url_on_click(modal, url_rect.x + url_rect.width / 2, url_rect.y + 1);
-        assert!(!opened, "click outside URL row should not open");
-        // Column outside the URL row's x-range.
-        let opened = state.maybe_open_url_on_click(
-            modal, modal.x, // left border column
-            url_rect.y,
-        );
-        assert!(!opened, "click on left border should not open");
-    }
-
-    #[test]
-    fn click_on_url_row_with_url_returns_true() {
-        let tmp = tempdir().unwrap();
-        let parent = tmp.path().join("parent");
-        let repo = parent.join("repo");
-        std::fs::create_dir_all(repo.join(".git")).unwrap();
-
-        let mut state = state_rooted_at(tmp.path().to_path_buf(), parent);
-        state.handle_key(key(KeyCode::Down));
-        state.handle_key(key(KeyCode::Enter));
-        state.pending_git_url = Some("file:///tmp/definitely-not-real".to_string());
-
-        let modal = manufactured_modal_area();
-        let url_rect = git_prompt_url_row_rect(modal, false).unwrap();
-        let opened =
-            state.maybe_open_url_on_click(modal, url_rect.x + url_rect.width / 2, url_rect.y);
-        assert!(opened, "click on URL row with URL should return true");
-        // Click doesn't dismiss the prompt.
-        assert!(state.pending_git_prompt.is_some());
-    }
-
-    // ── Entry name colour (WHITE) ─────────────────────────────────────
-
-    /// Plain (non-git) directory entries render their name in WHITE so
-    /// the listing stays legible against phosphor-green accents.
-    #[test]
-    fn non_git_entry_renders_in_white() {
-        use ratatui::{Terminal, backend::TestBackend};
-
-        let tmp = tempdir().unwrap();
-        std::fs::create_dir(tmp.path().join("plain")).unwrap();
-
-        let state = make_state_at(tmp.path().to_path_buf());
-        // Make sure nothing is selected so the highlight style doesn't
-        // mask the base WHITE colour we want to assert on.
-        let mut state = state;
-        state.list_state.select(None);
-
-        let backend = TestBackend::new(40, 10);
-        let mut terminal = Terminal::new(backend).unwrap();
-        terminal
-            .draw(|frame| {
-                render(frame, frame.area(), &state);
-            })
-            .unwrap();
-
-        let buffer = terminal.backend().buffer();
-        // Locate the first cell of the name "plain" — rows start at y=0
-        // with the block's top border, so the first entry sits at y=1
-        // and the name begins at x = 1 (border) + 2 (indent) = 3.
-        let cell = &buffer[(3, 1)];
-        assert_eq!(
-            cell.symbol(),
-            "p",
-            "expected 'p' at the entry's first char, got {:?}",
-            cell.symbol()
-        );
-        assert_eq!(
-            cell.fg,
-            Color::Rgb(255, 255, 255),
-            "non-git entry name should render in WHITE, got {:?}",
-            cell.fg
-        );
-    }
-
-    /// Git-repo entries render the name in WHITE and the ` (git)` suffix
-    /// in PHOSPHOR_GREEN so the marker pops against the otherwise-white row.
-    #[test]
-    fn git_entry_name_is_white_and_suffix_is_phosphor_green() {
-        use ratatui::{Terminal, backend::TestBackend};
-
-        let tmp = tempdir().unwrap();
-        let repo = tmp.path().join("repo");
-        std::fs::create_dir_all(repo.join(".git")).unwrap();
-
-        let mut state = make_state_at(tmp.path().to_path_buf());
-        // Clear selection so the highlight style doesn't mask the spans.
-        state.list_state.select(None);
-
-        let backend = TestBackend::new(40, 10);
-        let mut terminal = Terminal::new(backend).unwrap();
-        terminal
-            .draw(|frame| {
-                render(frame, frame.area(), &state);
-            })
-            .unwrap();
-
-        let buffer = terminal.backend().buffer();
-        // First entry row is at y=1 (below the block's top border).
-        // Name starts at x = 1 (border) + 2 (indent) = 3.
-        let name_cell = &buffer[(3, 1)];
-        assert_eq!(
-            name_cell.symbol(),
-            "r",
-            "expected 'r' at name's first char, got {:?}",
-            name_cell.symbol()
-        );
-        assert_eq!(
-            name_cell.fg,
-            Color::Rgb(255, 255, 255),
-            "git entry name should render in WHITE, got {:?}",
-            name_cell.fg
-        );
-
-        // Suffix: "  repo/ (git)" — the '(' of "(git)" sits at
-        // x = 3 (name start) + 5 (len of "repo/") + 1 (space) = 9.
-        let paren_cell = &buffer[(9, 1)];
-        assert_eq!(
-            paren_cell.symbol(),
-            "(",
-            "expected '(' at the suffix's first char, got {:?}",
-            paren_cell.symbol()
-        );
-        assert_eq!(
-            paren_cell.fg,
-            Color::Rgb(0, 255, 65),
-            "git suffix should render in PHOSPHOR_GREEN, got {:?}",
-            paren_cell.fg
-        );
-    }
-
-    // ── Symlink sandbox hardening ─────────────────────────────────────
-    //
-    // Finding #2 of the PR #166 current-branch review: lexical
-    // `Path::starts_with(root)` treated a symlink under `$HOME` as
-    // in-sandbox because its *path* starts with `$HOME`, but its
-    // canonical target could escape. Canonicalizing at the
-    // decision points fixes the leak.
-
-    #[cfg(unix)]
-    #[test]
-    fn symlink_to_outside_root_is_excluded_from_listing() {
-        let tmp = tempdir().unwrap();
-        let root = tmp.path().join("home");
-        let outside = tmp.path().join("outside");
-        std::fs::create_dir_all(&root).unwrap();
-        std::fs::create_dir_all(&outside).unwrap();
-        std::fs::create_dir_all(root.join("normal_dir")).unwrap();
-        // Symlink under root pointing at the sibling directory. A lexical
-        // `starts_with(root)` check accepts this path; a canonical one
-        // correctly rejects it.
-        std::os::unix::fs::symlink(&outside, root.join("escape_link")).unwrap();
-
-        let fb = FileBrowserState::new_at(root.clone(), root);
-        let names: Vec<&str> = fb.entries.iter().map(|e| e.name.as_str()).collect();
-        assert!(
-            names.contains(&"normal_dir"),
-            "regular child dir should still appear; got {names:?}",
-        );
-        assert!(
-            !names.contains(&"escape_link"),
-            "symlink escaping $HOME must not appear in listing; got {names:?}",
-        );
-    }
-
-    #[cfg(unix)]
-    #[test]
-    fn symlink_to_inside_root_still_appears() {
-        // Complementary test to `symlink_to_outside_root_is_excluded_from_listing`:
-        // we must not over-reject. A symlink that resolves back inside
-        // root is legitimate and should still be listed.
-        let tmp = tempdir().unwrap();
-        let root = tmp.path().join("home");
-        let inner = root.join("inner");
-        std::fs::create_dir_all(&inner).unwrap();
-        std::os::unix::fs::symlink(&inner, root.join("inner_link")).unwrap();
-
-        let fb = FileBrowserState::new_at(root.clone(), root);
-        let names: Vec<&str> = fb.entries.iter().map(|e| e.name.as_str()).collect();
-        assert!(names.contains(&"inner"));
-        assert!(
-            names.contains(&"inner_link"),
-            "symlink whose target stays inside root should still list; got {names:?}",
-        );
-    }
-
-    #[cfg(unix)]
-    #[test]
-    fn commit_rejects_out_of_root_target() {
-        // TOCTOU defence: even if an escaping path somehow reached
-        // `commit_or_reject` (list filtering beaten by a race, or a
-        // future bug elsewhere), the belt-and-suspenders check rejects it.
-        let tmp = tempdir().unwrap();
-        let root = tmp.path().join("home");
-        let outside = tmp.path().join("outside");
-        std::fs::create_dir_all(&root).unwrap();
-        std::fs::create_dir_all(&outside).unwrap();
-
-        let mut fb = FileBrowserState::new_at(root.clone(), root);
-        let outcome = fb.commit_or_reject(outside);
-        assert!(matches!(outcome, ModalOutcome::Continue));
-        let reason = fb
-            .rejected_reason
-            .as_deref()
-            .expect("out-of-root commit should set rejected_reason");
-        assert!(
-            reason.contains("outside"),
-            "rejection should cite the sandbox boundary; got {reason:?}",
-        );
-    }
-
-    #[test]
-    fn click_when_no_git_prompt_is_active_returns_false() {
-        let tmp = tempdir().unwrap();
-        let parent = tmp.path().join("parent");
-        std::fs::create_dir(&parent).unwrap();
-        let state = state_rooted_at(tmp.path().to_path_buf(), parent);
-        assert!(state.pending_git_prompt.is_none());
-
-        let modal = manufactured_modal_area();
-        let url_rect = git_prompt_url_row_rect(modal, false).unwrap();
-        let opened =
-            state.maybe_open_url_on_click(modal, url_rect.x + url_rect.width / 2, url_rect.y);
-        assert!(!opened, "click without active git prompt should be inert");
-    }
-}
